cleanup-incoming.pl: look into these 130.482 files what they have to say
git-svn-id: https://pause.perl.org:5460/svn/pause/trunk@180 a0804923-bdb8-0310-8f3c-e6a281b66435

+++ cron/cleanup-incoming.pl
@@ -27,6 +27,14 @@ for my $dirent (readdir DIR) {
   $sth->execute($dirent);
   next if $sth->rows > 0;
   my $size = -s $absdirent;
+  if ($dirent =~ /^(\d+)\.(\d+)$/) { # these come too often
+    open my $fh, $absdirent or die "Could not open $absdirent: $!";
+    local $/;
+    my $str = <$fh>;
+    substr($str,100*1024) = "" if length($str)> 100*1024;
+    require Data::Dumper;
+    warn sprintf "content[%s]\n", Data::Dumper::Dumper($str);
+  }
   unlink $absdirent or die "Could not unlink $absdirent: $!";
   warn "unlinked $absdirent ($size)\n";
 }
